pyproject.toml
[project]
name = "bmt-discord-bot"
version = "0.1.0"
description = "Discord bot for the Berkeley Math Tournament"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "asyncpg-stubs>=0.30.1",
    "discord-ext-menus",
    "discord-ext-menus-views",
    "discord-py>=2.5.2",
    "jishaku>=2.6.0",
    "parsedatetime>=2.6",
    "pillow>=11.2.1",
    "pylatex>=1.4.2",
    "pymupdf>=1.25.5",
    "python-dateutil>=2.9.0.post0",
    "typst>=0.13.2",
]

[tool.uv.sources]
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus" }
discord-ext-menus-views = { git = "https://github.com/oliver-ni/discord-ext-menus-views" }

[tool.ruff]
line-length = 100

[tool.ruff.lint.mccabe]
max-complexity = 100
